// src/app/models/config-form.model.ts
export interface ConfigFormModel {
	taxaSaque?: number;
	minimoSaque?: number;
	diasSaque?: number[];
	saqueInicio?: string;
	saqueFim?: string;
	diasSaqueBonus?: number[];
	saqueInicioBonus?: string;
	saqueFimBonus?: string;
	teto?: number;
	valorMinimo?: number;
	unilevel1?: number;
	unilevel2?: number;
	unilevel3?: number;
	unilevel4?: number;
	unilevel5?: number;
	unilevel6?: number;
	unilevel7?: number;
	unilevel8?: number;
	unilevel9?: number;
	unilevel10?: number;
	residual1?: number;
	residual2?: number;
	residual3?: number;
	residual4?: number;
	residual5?: number;
	residual6?: number;
	residual7?: number;
	residual8?: number;
	residual9?: number;
	residual10?: number;
	whatsapp?: string;
	whatsappGrupo?: string;
	rendimentoStake?: number;
	valorCripto?: number;
	banner1?: string | null;
	banner2?: string | null;
	banner3?: string | null;
}